aider: Updated `config.ts` to include an array of generated URLs from the `baseUrl` for the range 0 to 50.
# Aider chat conversation:

ASSISTANT: To make the requested change, we need to modify the `config.ts` file to update the `linkConfig.urls` property to be an array of generated URLs from the `baseUrl` for the range 0 to 50. Here's how to do it:

config.ts
<source>typescript
<<<<<<< SEARCH
linkConfig: {
  // urls: fill in here,
  regexps: [new RegExp('https://boardgamegeek.com/thread/\\d+/.*')],
  exclude: [new RegExp('https://boardgamegeek.com/thread/\\d+/article/.*')],
},
=======
linkConfig: {
  urls: Array.from({ length: 51 }, (_, i) => `${baseUrl}${i}`),
  regexps: [new RegExp('https://boardgamegeek.com/thread/\\d+/.*')],
  exclude: [new RegExp('https://boardgamegeek.com/thread/\\d+/article/.*')],
},
>>>>>>> REPLACE
</source>

This change will replace the comment `// urls: fill in here,` with an actual array of URLs generated from the `baseUrl` using a range from 0 to 50. Each URL will be created by appending the index `i` to the `baseUrl`.